feat: Docker containerization, remove Groq defaults, clean up dead code
- Add Dockerfile (multi-stage: frontend build, Python backend, nginx)
- Add docker-compose.yml (redis + web + worker + nginx)
- Add deploy/nginx-docker.conf for Docker deployment
- Remove /v2t/ URL prefix, deploy at root path /
- Remove Groq as default Whisper provider (user must configure WHISPER_BASE_URL)
- Remove state/ wrapper layer, callers use storage directly
- Delete memory_metadata.py and its test (dead code)
- Remove duplicate isStreaming computed from task store
- Rewrite README.md and CLAUDE.md to reflect current architecture
- Update .env.example, .dockerignore, Makefile
