+++ .dockerignore
@@ -7,17 +7,20 @@ __pycache__/
 *.py[cod]
 *$py.class
 build/
-dist/
 wheels/
 *.egg-info/
 .venv/
 
+# Node
+node_modules/
+frontend/dist/
+
 # Environment
 .env
 .env.*
+!.env.example
 
 # Testing
-tests/
 .pytest_cache/
 .coverage
 htmlcov/
@@ -32,14 +35,13 @@ htmlcov/
 .DS_Store
 Thumbs.db
 
-# Downloads
+# Downloads / data
 downloads/
 *.part
 
 # Docker
-Dockerfile
-docker-compose.yml
 .dockerignore
 
 # Documentation
 LICENSE
+*.md
