+++ backend/.env.example
@@ -1,29 +1,18 @@
-# OpenAI 兼容 API (AI 内容生成)
-OPENAI_API_KEY=your_api_key_here
+# AI 内容生成 (OpenAI 兼容)
+OPENAI_API_KEY=
 OPENAI_BASE_URL=https://api.deepseek.com
 OPENAI_MODEL=deepseek-reasoner
 
-# Whisper 兼容 API (语音转录)
-WHISPER_API_KEY=your_api_key_here
-WHISPER_BASE_URL=https://api.groq.com/openai/v1
-WHISPER_MODEL=whisper-large-v3
+# 语音转录 (OpenAI Whisper 兼容: Groq, OpenAI, 自托管 Qwen3-ASR 等)
+WHISPER_API_KEY=
+WHISPER_BASE_URL=
+WHISPER_MODEL=
 
-# 阿里云百炼 API (TTS 语音合成)
-DASHSCOPE_API_KEY=your_dashscope_api_key_here
+# 阿里云百炼 API (TTS 语音合成，可选)
+DASHSCOPE_API_KEY=
 
-# Xiazaitool API (备用下载，可选)
-XIAZAITOOL_TOKEN=your_xiazaitool_token_here
+# Xiazaitool API (视频链接解析)
+XIAZAITOOL_TOKEN=
 
-# 下载目录
-DOWNLOAD_DIR=./downloads
-
-# MongoDB
-MONGODB_URI=mongodb://localhost:27017
-MONGODB_DATABASE=v2t
-
-# MinIO
-MINIO_ENDPOINT=localhost:9000
-MINIO_ACCESS_KEY=minioadmin
-MINIO_SECRET_KEY=minioadmin
-MINIO_BUCKET=v2t
-MINIO_SECURE=false
+# Redis
+REDIS_URL=redis://localhost:6379
