+++ backend/app/routers/stream.py
@@ -19,8 +19,7 @@ from app.services.llm import (
     generate_zhihu_article,
 )
 from app.services.podcast_tts import PodcastTTSError, generate_podcast_audio
-from app.state import get_workspace, save_workspace
-from app.storage import get_file_storage
+from app.storage import get_file_storage, get_workspace, save_workspace
 from app.utils.sse import sse_data, sse_response
 
 logger = logging.getLogger(__name__)
