+++ backend/app/state/__init__.py
@@ -1,10 +1,6 @@
-"""状态管理模块"""
+"""状态管理模块 — 已迁移到 app.storage，保留兼容导出"""
 
-from .memory_store import (
-    get_workspace,
-    register_workspace,
-    save_workspace,
-)
+from app.storage import get_workspace, register_workspace, save_workspace
 
 __all__ = [
     "get_workspace",
