+++ backend/app/tasks/workspace_task.py
@@ -8,8 +8,7 @@ from app.models.entities import Workspace, WorkspaceResource
 from app.models.enums import ResourceType, WorkspaceStatus
 from app.services.transcribe import TranscribeError, extract_audio_async, transcribe_audio
 from app.services.video_downloader import DownloadError, download_video
-from app.state import get_workspace, save_workspace
-from app.storage import get_file_storage, get_redis
+from app.storage import get_file_storage, get_redis, get_workspace, save_workspace
 from app.utils.response import build_workspace_response
 
 logger = logging.getLogger(__name__)
