+++ deploy/nginx-docker.conf
@@ -1,0 +1,20 @@
+server {
+    listen 80;
+
+    # 前端 SPA
+    location / {
+        root /usr/share/nginx/html;
+        try_files $uri $uri/ /index.html;
+    }
+
+    # API 反代
+    location /api/ {
+        proxy_pass http://web:8101;
+        proxy_read_timeout 300s;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header Connection '';
+        proxy_http_version 1.1;
+        proxy_buffering off;
+    }
+}
