+++ deploy/v2t.service
@@ -6,10 +6,7 @@ After=network.target redis.service
 Type=simple
 User=YOUR_USER
 WorkingDirectory=/path/to/v2t/backend
-# 如果部署在根路径，使用这行：
-# ExecStart=.venv/bin/uvicorn app.main:app --host 0.0.0.0 --port 8101
-# 如果部署在 /v2t 子路径，使用这行：
-ExecStart=.venv/bin/uvicorn app.main:app --host 0.0.0.0 --port 8101 --root-path /v2t
+ExecStart=.venv/bin/uvicorn app.main:app --host 0.0.0.0 --port 8101
 Restart=always
 RestartSec=10
 Environment=PATH=.venv/bin:/usr/local/bin:/usr/bin:/bin
