+++ docker-compose.yml
@@ -1,0 +1,50 @@
+services:
+  redis:
+    image: redis:7-alpine
+    volumes:
+      - redis-data:/data
+    restart: unless-stopped
+
+  web:
+    build:
+      context: .
+      target: backend
+    depends_on:
+      - redis
+    env_file: .env
+    environment:
+      REDIS_URL: redis://redis:6379
+    volumes:
+      - v2t-data:/data/v2t
+      - v2t-tmp:/tmp/v2t
+    restart: unless-stopped
+
+  worker:
+    build:
+      context: .
+      target: backend
+    command: [".venv/bin/arq", "app.worker.WorkerSettings"]
+    depends_on:
+      - redis
+    env_file: .env
+    environment:
+      REDIS_URL: redis://redis:6379
+    volumes:
+      - v2t-data:/data/v2t
+      - v2t-tmp:/tmp/v2t
+    restart: unless-stopped
+
+  nginx:
+    build:
+      context: .
+      target: nginx
+    ports:
+      - "8100:80"
+    depends_on:
+      - web
+    restart: unless-stopped
+
+volumes:
+  redis-data:
+  v2t-data:
+  v2t-tmp:
